requirements.txt
Flask==2.3.2
keras==2.12.0
Pillow==9.5.0
tensorflow==2.12.0
torch==2.0.1
